requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
black

# machine learning packages
numpy
scipy
scikit-learn
pandas
matplotlib
seaborn